'已完成提交'

+++ normal_biz/src/main/java/com/normal/biz/order/OrderMapper.java
@@ -10,6 +10,6 @@ public interface OrderMapper {
 
     Order selectByPrimaryKey(Integer id);
 
-    int updateOrderStatus(@Param("id") int id, @Param("orderStatus") OrderStatus orderStatus);
+    int updateOrderStatus(@Param("id") long id, @Param("orderStatus") OrderStatus orderStatus);
 
 }
